Added events.js
feat : removed the events data from components and added as separate file(data/event.js) .
