+++ src/components/Events.jsx
@@ -1,80 +1,7 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import Snow from '../ui/Snow';
-
-// Event data with all necessary details
-const eventsData = [
-  {
-    id: 1,
-    title: "Hawkin's High Papers",
-    slug: "hawkins-high-papers",
-    description: "Paper Presentation event focusing on academic writing and research works.",
-    image: "/events/hole.png",
-    rules: `
-Rules:
-  1. Maximum number of participants in a team is 3.
-  2. Only Engineering students are allowed.
-  3. Bring your College ID and Bonafide Certificate.
-  4. Submitted papers must be in IEEE format.
-  5. Papers must not be published before anywhere or be under consideration for publication in any journal or conference.
-  6. Each team presents their paper for 7 minutes followed by a Q&A session for 3 minutes.
-  7. Plagiarism is not allowed (above 40%). Original works only.
-  8. Abstract should not exceed more than 250 words.
-  9. Paper should have a minimum of 6 pages and should not exceed 15 pages.
-  10. Soft-copies of the submitted paper must be in standard document formats (docx, doc, pdf).
-  11. Paper must be preceded by a cover page specifying the title of the paper, names of authors and their college names as also their contact numbers and email ids
-  12. Persons from different institutions can be a part of the same team. However, one person may not be a part of multiple teams for the same event.
-  13. The paper is to be attached to the mail flairit@licet.ac.in with the subject as the title of the paper.
-  14. The mail must contain - name of authors, college names and departments, phone numbers, and email addresses. 
-    `,
-    date: "16th September 2023",
-    time: "10:00 am to 11:30 am",
-    location: "Case Tools Lab (H23) and CIS Lab (J14)",
-    organizers: "Ashik S and Rithika R (Contact No: 9445387601)",
-    registerationlink: "https://docs.google.com/forms/d/e/1FAIpQLSfrG7uSLaSLP3EHutd59-OCYY4SsMLoj0X-O1Eiwgp0ppo6Ew/viewform?usp=sf_link"
-  },
-  { 
-    id: 2,
-    title: "Arcade of Answers",
-    slug: "arcade-of-answers",
-    description: "Technical quiz challenging your knowledge across various domains.",
-    image: "/events/ray.png",
-    rules: `
-Rules:
-  1. Teams of 2-3 members
-  2. Multiple choice questions across various tech domains
-  3. No external resources allowed
-  4. Three rounds: Elimination, Rapid Fire, and Final Showdown
-  5. Decisions of quiz masters are final
-    `,
-    date: "16th September 2023",
-    time: "1:30 pm to 3:00 pm",
-    location: "RDBMS Laboratory (A21)",
-    organizers: "Regis R and Arul Danica (Contact: 9150206627)",
-    registerationlink: "https://docs.google.com/forms/d/e/1FAIpQLSd_DfIN1W9c-YCo04bByXb6rmDJahOAm4K1p7z15ypjFygAEg/viewform?usp=sf_link"
-  },
-  // Add other events with similar structure
-  {
-    id: 8,
-    title: "The Upside Down",
-    slug: "the-upside-down",
-    description: "Cyber ciphers challenge to solve encrypted puzzles.",
-    image: "/events/house.png",
-    rules: `
-Rules:
-  1. Individual participation
-  2. Three rounds of increasing difficulty
-  3. No internet access allowed
-  4. Time-bound challenges
-  5. Winners decided by fastest correct solutions
-    `,
-    date: "16th September 2023",
-    time: "1:30 pm to 3:00 pm",
-    location: "Classroom (I32)",
-    organizers: "Nivedita Ramatoti and Chrispin Sheena (Contact: 7418921212)",
-    registerationlink: "https://docs.google.com/forms/d/e/1FAIpQLSc0WFuedTBhq-WCORqNML5eraM5zMO662WWVirtcMnvh-Ttbg/viewform?usp=sf_link"
-  }
-];
+import Snow from '../ui/Smoke';
+import eventsData from '../data/events';
 
 const Events = () => {
   const navigate = useNavigate();
